pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = 'checkmate'
version = '0.1.3'
description = 'An API for testing a Python function on an input / output suite.'
authors = [
    {name = 'Alkis Gotovos', email = 'alkis.gotovos@gmail.com'},
]
license = 'MIT'
requires-python = '>=3.9'
dependencies = [
    'pydantic==1.10.7',
]

[project.optional-dependencies]
test = [
    'fastapi>=0.100.1',
    'httpx>=0.25.0',
    'pytest-cov',
]
api = [
    'fastapi>=0.100.1',
    'httpx>=0.25.0',
]

[tool.black]
line-length = 119